+++ src/org/camunda/latera/bss/connectors/hid/Hydra.groovy
@@ -36,11 +36,13 @@ import org.camunda.latera.bss.connectors.hid.hydra.Tag
 import org.camunda.latera.bss.internal.Version
 
 class Hydra implements Ref, Message, DataType, AddParam, Good, Document, Contract, PriceOrder, Invoice, Bill, Subject, Company, Person, Reseller, Group, Customer, Account, Subscription, Equipment, Region, Address, Param, Search, Tag {
+  private static String DEFAULT_USER   = 'hydra'
+  private static String DEFAULT_LOCALE = 'ru'
   HID hid
   String user
   private String password
-  def firmId
-  def resellerId
+  Number firmId
+  Number resellerId
   SimpleLogger logger
   String locale
   Version version
@@ -51,10 +53,10 @@ class Hydra implements Ref, Message, DataType, AddParam, Good, Document, Contrac
     this.hid        = new HID(execution)
     def ENV         = System.getenv()
 
-    this.locale     = execution.getVariable('locale')
-    this.user       = ENV['HYDRA_USER']     ?: execution.getVariable('hydraUser') ?: 'hydra'
+    this.locale     = execution.getVariable('locale') ?: DEFAULT_LOCALE
+    this.user       = ENV['HYDRA_USER']     ?: execution.getVariable('hydraUser') ?: DEFAULT_USER
     this.password   = ENV['HYDRA_PASSWORD'] ?: execution.getVariable('hydraPassword')
-    this.firmId     = toIntSafe(execution.getVariable('hydraFirmId')     ?: (execution.getVariable('homsOrderDataFirmId') ?: getDefaultFirmId()))
+    this.firmId     = toIntSafe(execution.getVariable('hydraFirmId')     ?: (execution.getVariable('homsOrderDataFirmId') ?: DEFAULT_FIRM))
     this.resellerId = toIntSafe(execution.getVariable('hydraResellerId') ?: execution.getVariable('homsOrderDataResellerId'))
     this.regionHierarchyOverride = execution.getVariable('regionHierarchy')
 
@@ -107,24 +109,7 @@ class Hydra implements Ref, Message, DataType, AddParam, Good, Document, Contrac
     return result
   }
 
-  Number getDefaultFirmId() {
-    return DEFAULT_FIRM
-  }
-
-  Number getFirmId() {
-    return firmId
-  }
-
-  Number getResellerId() {
-    return resellerId
-  }
-
-  String getLocale() {
-    return locale ?: 'ru'
-  }
-
   Number getLangId() {
-    String locale = getLocale()
     return getRefIdByCode("LANG_${capitalize(locale)}")
   }